# verity-server/spring-boot-server/springServer/src/main/resources/verityCore-swagger20.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.1",
        "title": "VerityReputation",
...
    },
    "host": "localhost:8080",
    "basePath": "/verity/core",
    "schemes": [
        "http",
        "https"
    ],
    "securityDefinitions": {
        "Basic Auth": {
            "type": "basic",
            "description": "Basic Authentication for now while we build and test the api"
        },
        "verity_auth_token": {
            "name": "verity_auth_token",
            "type": "apiKey",
            "in": "query"
        }
    },
    "paths": {
...
        "/person": {
            "post": {
                "operationId": "createPerson",
                "summary": "create person",
                "tags": [
                    "Person"
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/Person"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "muilti-hash blockchain id for the newly created person account",
                        "schema": {
                            "$ref": "#/definitions/Person"
                        }
                    },
                    "401": {
                        "$ref": "#/responses/trait:standardErrors:401"
                    },
                    "403": {
                        "$ref": "#/responses/trait:standardErrors:403"
                    },
                    "404": {
                        "$ref": "#/responses/trait:standardErrors:404"
                    },
                    "409": {
                        "$ref": "#/responses/trait:standardErrors:409"
                    },
                    "422": {
                        "$ref": "#/responses/trait:standardErrors:422"
                    },
                    "500": {
                        "$ref": "#/responses/trait:standardErrors:500"
                    }
                },
                "x-stoplight": {
                    "id": "createPerson",
                    "beforeScript": null,
                    "afterScript": null,
                    "public": true,
                    "mock": {
                        "statusCode": 200,
                        "enabled": false
                    }
                }
            },
            "put": {
                "operationId": "updatePerson",
                "summary": "update person",
                "tags": [
                    "Person"
                ],
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/Person"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Status 200"
                    },
                    "401": {
                        "$ref": "#/responses/trait:standardErrors:401"
                    },
                    "403": {
                        "$ref": "#/responses/trait:standardErrors:403"
                    },
                    "404": {
                        "$ref": "#/responses/trait:standardErrors:404"
                    },
                    "409": {
                        "$ref": "#/responses/trait:standardErrors:409"
                    },
                    "422": {
                        "$ref": "#/responses/trait:standardErrors:422"
                    },
                    "500": {
                        "$ref": "#/responses/trait:standardErrors:500"
                    }
                },
                "x-stoplight": {
                    "id": "updatePerson",
                    "beforeScript": null,
                    "afterScript": null,
                    "public": true,
                    "mock": {
                        "statusCode": 200,
                        "enabled": false
                    }
                }
            }
        },
        "/person/{uuid}": {
            "parameters": [
                {
                    "name": "uuid",
                    "in": "path",
                    "description": "multi-hash id of person record on the blockchain",
                    "required": true,
                    "type": "string"
                }
            ],
            "get": {
                "operationId": "getPerson",
                "summary": "get Person",
                "tags": [
                    "Person"
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Status 200",
                        "schema": {
                            "$ref": "#/definitions/Person"
                        }
                    }
                },
                "x-stoplight": {
                    "id": "getPerson",
                    "beforeScript": null,
                    "afterScript": null,
                    "public": true,
                    "mock": {
                        "statusCode": 200,
                        "enabled": false
                    }
                }
            }
        },
        "/person/{uuid}/content": {
            "parameters": [
                {
                    "name": "uuid",
                    "in": "path",
                    "description": "multi-hash id of person record on the blockchain",
                    "required": true,
                    "type": "string"
                }
            ],
            "get": {
                "operationId": "getPersonContent",
                "summary": "get Person Content",
                "tags": [
                    "Person"
                ],
                "description": "Get all the content this person has authored or co-authored",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "$ref": "#/parameters/trait:byDateRange:startDate"
                    },
                    {
                        "$ref": "#/parameters/trait:byDateRange:endDate"
                    },
                    {
                        "$ref": "#/parameters/trait:paged:pageNumber"
                    },
                    {
                        "$ref": "#/parameters/trait:paged:pageSize"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Status 200",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Content"
                            }
                        }
                    }
                },
                "x-stoplight": {
                    "id": "getPersonContent",
                    "beforeScript": null,
                    "afterScript": null,
                    "public": true,
                    "mock": {
                        "statusCode": 200,
                        "enabled": false
                    }
                }
            }
        },
        "/person/{uuid}/reputation/{communityid}/{valuetypeid}/{reputationtype}": {
            "parameters": [
                {
                    "name": "uuid",
                    "in": "path",
                    "required": true,
                    "type": "string"
                },
                {
                    "name": "communityid",
                    "in": "path",
                    "required": true,
                    "type": "string"
                },
                {
                    "name": "valuetypeid",
                    "in": "path",
                    "required": true,
                    "type": "string"
                },
                {
                    "name": "reputationtype",
                    "in": "path",
                    "required": true,
                    "type": "string"
                }
            ],
            "get": {
                "operationId": "GET_person-uuid-reputation-communityid-valuetypeid-reputationtype",
                "summary": "get person reputation",
                "tags": [
                    "Person"
                ],
                "description": "Get the reputation as calculated by the algorithm requested in the reputationtype parameter.",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "schema": {
                            "type": "number",
                            "description": "the result of the reputation calculation as a numerical value"
                        }
                    }
                },
                "x-stoplight": {
                    "id": "GET_person-uuid-reputation-communityid-valuetypeid-reputationtype",
                    "beforeScript": null,
                    "afterScript": null,
                    "public": true,
                    "mock": {
                        "enabled": false,
                        "dynamic": false,
                        "statusCode": 200
                    }
                }
            }
        },
...
    },
    "parameters": {
...
        "trait:paged:pageNumber": {
            "name": "pageNumber",
            "in": "query",
            "description": "starting page number. Must be > 0.",
            "required": true,
            "type": "integer",
            "minimum": 1
        },
        "trait:paged:pageSize": {
            "name": "pageSize",
            "in": "query",
            "description": "number of records to return per page.",
            "required": true,
            "type": "integer",
            "minimum": 1,
            "maximum": 1000
        },
        "trait:byDateRange:startDate": {
            "name": "startDate",
            "in": "query",
            "required": true,
            "type": "string"
        },
        "trait:byDateRange:endDate": {
            "name": "endDate",
            "in": "query",
            "required": true,
            "type": "string"
        }
    },
    "responses": {
        "trait:standardErrors:401": {
            "description": "",
            "schema": {
                "type": "object"
            }
        },
        "trait:standardErrors:403": {
            "description": "",
            "schema": {
                "type": "object",
                "properties": {
                    "message": {
                        "type": "string"
                    }
                },
                "required": [
                    "message"
                ]
            }
        },
        "trait:standardErrors:404": {
            "description": "",
            "schema": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "string"
                    },
                    "error": {
                        "type": "string"
                    }
                },
                "required": [
                    "status",
                    "error"
                ]
            }
        },
        "trait:standardErrors:409": {
            "description": "",
            "schema": {
                "type": "object",
                "description": "409 Conflict. https://tools.ietf.org/html/rfc7231#section-6.5.8"
            }
        },
        "trait:standardErrors:422": {
            "description": "",
            "schema": {
                "type": "object",
                "description": "422 Unprocessable Entity. http://tools.ietf.org/html/rfc4918#section-11.2"
            }
        },
        "trait:standardErrors:500": {
            "description": "",
            "schema": {
                "type": "object"
            }
        }
    },
    "definitions": {
...
        "Person": {
            "title": "Person",
            "type": "object",
            "properties": {
                "agent": {
                    "$ref": "#/definitions/Agent"
                },
                "uuid": {
                    "type": "string",
                    "description": "UUID, GUID, HASH,  MultiHash or ProxyContract Address that represents this object"
                },
                "firstName": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "nickName": {
                    "type": "string"
                },
                "organizationId": {
                    "type": "string"
                },
                "description": {
                    "$ref": "#/definitions/Content"
                }
            },
            "required": [
                "agent",
                "uuid",
                "nickName"
            ],
            "x-stoplight": {
                "id": "Person",
                "name": "Person",
                "public": true
            }
        },
...
        "Content": {
            "title": "Content",
            "type": "object",
            "description": "Self describing content (document object model, JSON-LD) A physical, digital, conceptual, or other kind of thing with some fixed aspects; entities may be real or imaginary.\n\nMaps to https://www.w3.org/TR/prov-o/  prov:Entity",
            "properties": {
                "uuid": {
                    "type": "string",
                    "description": "UUID, GUID, HASH or MultiHash that represents this object. For example a GitHub commit hash or IPFS address."
                },
                "name": {
                    "type": "string",
                    "description": "name or tag"
                },
                "body": {
                    "type": "string",
                    "description": "content as text or JSON"
                },
                "nodes": {
                    "type": "array",
                    "description": "child node ids",
                    "uniqueItems": true,
                    "items": {
                        "type": "string"
                    }
                },
                "authors": {
                    "type": "array",
                    "description": "one or more uuids of the creating agent(s)",
                    "items": {
                        "type": "string"
                    }
                },
                "communityId": {
                    "type": "string",
                    "description": "The community from which the content originated."
                }
            },
            "required": [
                "uuid",
                "name",
                "body"
            ],
            "x-stoplight": {
                "id": "Content",
                "name": "Content",
                "public": true
            }
        },
        "Agent": {
            "title": "Agent",
            "type": "object",
            "description": "Agent is something that bears some form of responsibility for changing state in the Verity system. An Agent could be a smart contract, person, organization, device, or anything the owns the Private Key half of the publicKey of the Agent and controlls the accounts used for attestationURLs.\n\nMaps to  https://www.w3.org/TR/prov-o/#Agent  prov:Agent",
            "properties": {
                "uuid": {
                    "type": "string",
                    "description": "UUID, GUID, HASH,  MultiHash or ProxyContract Address that represents this object"
                },
                "publicKey": {
                    "type": "string"
                },
                "attestationUrls": {
                    "description": "zero or more attestation URLs (twitter, FB, blog) links to proof of account ownership, a post with public key id of owner. Signing the attribute with the\nVerity signing key creates a claim that the Verity system has created this attestation. \n\nExamples at: \nhttps://keybase.io/\nhttps://keybase.io/docs/kbfs/understanding_kbfs\n\nHow to use Oracalize to create attestations: http://docs.oraclize.it/#datasources-url.\nAlignment with W3C Verifiable Claims spec is encouraged. https://opencreds.github.io/vc-data-model/",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            },
            "required": [
                "uuid",
                "publicKey"
            ],
            "x-stoplight": {
                "id": "Agent",
                "name": "Agent",
                "public": true
            }
        },
...
    },
...
}